move find_project_root to common helper func, refactor it into relevant scripts
